# Optimize memoization helpers.